Extracted Copyright component to file
