+++ client/src/components/Copyright.js
@@ -1,0 +1,17 @@
+import { Link, Typography } from '@material-ui/core';
+import React from 'react';
+
+function Copyright() {
+    return (
+      <Typography variant="body2" color="textSecondary" align="center">
+        {"Copyright © "}
+        <Link color="inherit" href="#">
+          LovingSitter
+        </Link>{" "}
+        {new Date().getFullYear()}
+        {"."}
+      </Typography>
+    );
+}
+
+export default Copyright;
