+++ client/src/pages/SignUp.js
@@ -12,19 +12,7 @@ import useForm from '../components/useForm';
 import Input from '../components/controls/Input';
 import { Toolbar } from '@material-ui/core';
 
-
-function Copyright() {
-  return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {'Copyright © '}
-      <Link color="inherit" href="#">
-        LovingSitter
-      </Link>{' '}
-      {new Date().getFullYear()}
-      {'.'}
-    </Typography>
-  );
-}
+import Copyright from '../components/Copyright';
 
 const useStyles = makeStyles((theme) => ({
   paper: {
